refactor: extract internal classes and keyboard shortcuts from RichTextEditor to separate files
