+++ richtext-editor/src/main/kotlin/dev/mkeeda/arranger/richtext/editor/RichTextInputTransformation.kt
@@ -1,0 +1,12 @@
+package dev.mkeeda.arranger.richtext.editor
+
+import androidx.compose.foundation.text.input.InputTransformation
+import androidx.compose.foundation.text.input.TextFieldBuffer
+
+internal class RichTextInputTransformation(
+    private val state: RichTextState,
+) : InputTransformation {
+    override fun TextFieldBuffer.transformInput() {
+        state.updateRichString(this)
+    }
+}
